+++ src/orbiter/src/http/routes/api/not_allowed.rs
@@ -1,0 +1,39 @@
+use crate::http::routes::api::init::init_certified_response;
+use crate::http::routes::api::routes::{
+    EVENTS_ROUTE, EVENT_ROUTE, METRICS_ROUTE, METRIC_ROUTE, VIEWS_ROUTE, VIEW_ROUTE,
+};
+use crate::http::routes::api::types::CertifiedExactRoute;
+use crate::http::types::interface::ErrorResponse;
+use crate::http::utils::create_json_response;
+use ic_http_certification::{HttpResponse, Method, StatusCode};
+
+pub fn init_certified_not_allowed_responses() {
+    init_not_allowed_responses(&VIEW_ROUTE);
+    init_not_allowed_responses(&VIEWS_ROUTE);
+    init_not_allowed_responses(&EVENT_ROUTE);
+    init_not_allowed_responses(&EVENTS_ROUTE);
+    init_not_allowed_responses(&METRIC_ROUTE);
+    init_not_allowed_responses(&METRICS_ROUTE);
+}
+
+fn init_not_allowed_responses(certified_route: &CertifiedExactRoute) {
+    [
+        Method::GET,
+        Method::DELETE,
+        Method::HEAD,
+        Method::PUT,
+        Method::PATCH,
+        Method::OPTIONS,
+        Method::TRACE,
+        Method::CONNECT,
+    ]
+    .into_iter()
+    .for_each(|method| {
+        fn create_not_allowed_response() -> HttpResponse<'static> {
+            let body = ErrorResponse::not_allowed().encode();
+            create_json_response(StatusCode::METHOD_NOT_ALLOWED, body)
+        }
+
+        init_certified_response(certified_route, method, create_not_allowed_response);
+    });
+}
