+++ src/orbiter/src/http/routes/mod.rs
@@ -1,5 +1,3 @@
-pub mod not_allowed;
+pub mod api;
 pub mod not_found;
 mod services;
-mod setup;
-mod types;
