+++ src/orbiter/src/http/server.rs
@@ -1,5 +1,5 @@
 use crate::http::constants::NOT_FOUND_PATH;
-use crate::http::routes::not_allowed::prepare_certified_not_allowed_response;
+use crate::http::routes::api::services::prepare_certified_not_allowed_response;
 use crate::http::routes::not_found::{
     create_uncertified_not_found_response, prepare_certified_not_found_response,
 };
